pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "flowc"
version = "0.1.0"
description = "CLI for managing GitHub activity across configured repos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["python-dotenv"]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
flowc = "fc_cli.main:main"

[tool.setuptools.packages.find]
include = ["fc_cli*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests_fc"]
